Updated to do Spirit Weapons and multiple prices
Spirit Weapons have a tabbed Base Stats and Information table, one for normal
and one for the Spirit Weapon version, which was breaking everything. Made some
major changes to fix that. Ended up getting rid of CSSSelect entirely; XPaths
for needed elements are now hardcoded in. Also added support for multiple
prices each with lists of sellers in the Sold By.