The final script (kinda)
Cleaned up the code in a lot of ways. Made exception handling more robust
and as specific as possible, and all progress messages, WARNINGs, and ERRORs
print to stderr now. Removed CRITICALs and converted those to normal ERRORs,
so the datafile should never be corrupted. The command to run is in a comment
at the bottom of the file.