Changed table XPath for more reliability
Just changed the position parameter on following-sibling:: to match one or two
away, lol. This fixes the problem with pages that have an image between the
Base Stats and Information header and the table. Can't go any higher than this,
though, or risk selecting other tables. Most likely, this will fix the majority
of problematic (ERROR, not WARNING) pages.